Step 5d: Refactor hooks
